refactoring: just the owner can delete a thread

+++ app/Controllers/Http/ThreadController.js
@@ -4,16 +4,17 @@ const Thread = use("App/Models/Thread");
 
 class ThreadController {
   async store({ request, response, auth }) {
-    const attributes = {
-      ...request.only(["title", "body"]),
-      user_id: auth.user.id,
-    };
-    const thread = await Thread.create(attributes);
+    const thread = await auth.user
+      .threads()
+      .create(request.only(["title", "body"]));
     return response.json({ thread });
   }
 
-  async destroy({ params }) {
+  async destroy({ params, response, auth }) {
     const thread = await Thread.findOrFail(params.id);
+    if (thread.user_id !== auth.user.id) {
+      return response.status(403).send();
+    }
     await thread.delete();
   }
 }
